requirements.txt
blinker==1.4
boto3==1.19.3
botocore==1.22.3
click==8.0.3
Flask==2.0.2
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.1
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.26
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==2.3.3
